Backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
django-extensions==3.2.3
python-decouple==3.8

djangorestframework-simplejwt==5.4.0
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0

cloudinary==1.41.0
django-cloudinary-storage==0.3.0

psycopg2-binary==2.9.9
